requirements.txt
playwright==1.51
openai
pillow
tenacity
pyyaml
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
pytest>=7.0
pytest-asyncio>=0.21

